flip build switch to AMD
