+++ package.json
@@ -11,17 +11,17 @@
   "scripts": {
     "test": "echo Please run any of the test scripts from the scripts folder.",
     "test-browser": "npx playwright install && node test/unit/browser/index.js",
-    "test-browser-esm": "npx playwright install && node test/unit/browser/index.esm.js",
+    "test-browser-amd": "npx playwright install && node test/unit/browser/index.amd.js",
     "test-browser-no-install": "node test/unit/browser/index.js",
-    "test-browser-esm-no-install": "node test/unit/browser/index.esm.js",
+    "test-browser-amd-no-install": "node test/unit/browser/index.amd.js",
     "test-node": "mocha test/unit/node/index.js --delay --ui=tdd --timeout=5000 --exit",
-    "test-node-esm": "mocha test/unit/node/index.mjs --delay --ui=tdd --timeout=5000 --exit",
+    "test-node-amd": "mocha test/unit/node/index-amd.js --delay --ui=tdd --timeout=5000 --exit",
     "test-extension": "vscode-test",
     "preinstall": "node build/npm/preinstall.js",
     "postinstall": "node build/npm/postinstall.js",
     "compile": "node --max-old-space-size=4095 ./node_modules/gulp/bin/gulp.js compile",
     "watch": "npm-run-all -lp watch-client watch-extensions",
-    "watch-esm": "npm-run-all -lp watch-client-esm watch-extensions",
+    "watch-amd": "npm-run-all -lp watch-client-amd watch-extensions",
     "watchd": "deemon yarn watch",
     "watch-webd": "deemon yarn watch-web",
     "kill-watchd": "deemon --kill yarn watch",
@@ -29,7 +29,7 @@
     "restart-watchd": "deemon --restart yarn watch",
     "restart-watch-webd": "deemon --restart yarn watch-web",
     "watch-client": "node --max-old-space-size=4095 ./node_modules/gulp/bin/gulp.js watch-client",
-    "watch-client-esm": "node --max-old-space-size=4095 ./node_modules/gulp/bin/gulp.js watch-client-esm",
+    "watch-client-amd": "node --max-old-space-size=4095 ./node_modules/gulp/bin/gulp.js watch-client-amd",
     "watch-clientd": "deemon yarn watch-client",
     "kill-watch-clientd": "deemon --kill yarn watch-client",
     "watch-extensions": "node --max-old-space-size=4095 ./node_modules/gulp/bin/gulp.js watch-extensions watch-extension-media",
