+++ test/unit/browser/index.amd.js
@@ -10,7 +10,7 @@ const path = require('path');
 const glob = require('glob');
 const events = require('events');
 const mocha = require('mocha');
-const createStatsCollector = require('../../../node_modules/mocha/lib/stats-collector');
+const createStatsCollector = require('mocha/lib/stats-collector');
 const MochaJUnitReporter = require('mocha-junit-reporter');
 const url = require('url');
 const minimatch = require('minimatch');
@@ -21,7 +21,6 @@ const yaserver = require('yaserver');
 const http = require('http');
 const { randomBytes } = require('crypto');
 const minimist = require('minimist');
-const { promisify } = require('node:util');
 
 /**
  * @type {{
@@ -83,8 +82,6 @@ Options:
 --help, -h           show the help`);
 	process.exit(0);
 }
-
-const isDebug = !!args.debug;
 
 const withReporter = (function () {
 	if (args.tfs) {
@@ -186,14 +183,10 @@ async function createServer() {
 			req.on('end', () => resolve(JSON.parse(Buffer.concat(body).toString())));
 			req.on('error', reject);
 		});
-		try {
-			const result = await fn(...params);
-			response.writeHead(200, { 'Content-Type': 'application/json' });
-			response.end(JSON.stringify(result));
-		} catch (err) {
-			response.writeHead(500);
-			response.end(err.message);
-		}
+
+		const result = await fn(...params);
+		response.writeHead(200, { 'Content-Type': 'application/json' });
+		response.end(JSON.stringify(result));
 	};
 
 	const server = http.createServer((request, response) => {
@@ -204,24 +197,17 @@ async function createServer() {
 		// rewrite the URL so the static server can handle the request correctly
 		request.url = request.url.slice(prefix.length);
 
-		function massagePath(p) {
-			// TODO@jrieken FISHY but it enables snapshot
-			// in ESM browser tests
-			p = String(p).replace(/\\/g, '/').replace(prefix, rootDir);
-			return p;
-		}
-
 		switch (request.url) {
 			case '/remoteMethod/__readFileInTests':
-				return remoteMethod(request, response, p => fs.promises.readFile(massagePath(p), 'utf-8'));
+				return remoteMethod(request, response, p => fs.promises.readFile(p, 'utf-8'));
 			case '/remoteMethod/__writeFileInTests':
-				return remoteMethod(request, response, (p, contents) => fs.promises.writeFile(massagePath(p), contents));
+				return remoteMethod(request, response, (p, contents) => fs.promises.writeFile(p, contents));
 			case '/remoteMethod/__readDirInTests':
-				return remoteMethod(request, response, p => fs.promises.readdir(massagePath(p)));
+				return remoteMethod(request, response, p => fs.promises.readdir(p));
 			case '/remoteMethod/__unlinkInTests':
-				return remoteMethod(request, response, p => fs.promises.unlink(massagePath(p)));
+				return remoteMethod(request, response, p => fs.promises.unlink(p));
 			case '/remoteMethod/__mkdirPInTests':
-				return remoteMethod(request, response, p => fs.promises.mkdir(massagePath(p), { recursive: true }));
+				return remoteMethod(request, response, p => fs.promises.mkdir(p, { recursive: true }));
 			default:
 				return serveStatic.handle(request, response);
 		}
@@ -244,7 +230,7 @@ async function runTestsInBrowser(testModules, browserType) {
 	const browser = await playwright[browserType].launch({ headless: !Boolean(args.debug), devtools: Boolean(args.debug) });
 	const context = await browser.newContext();
 	const page = await context.newPage();
-	const target = new URL(server.url + '/test/unit/browser/renderer.esm.html');
+	const target = new URL(server.url + '/test/unit/browser/renderer.amd.html');
 	target.searchParams.set('baseUrl', url.pathToFileURL(path.join(rootDir, 'src')).toString());
 	if (args.build) {
 		target.searchParams.set('build', 'true');
@@ -252,12 +238,6 @@ async function runTestsInBrowser(testModules, browserType) {
 	if (process.env.BUILD_ARTIFACTSTAGINGDIRECTORY) {
 		target.searchParams.set('ci', 'true');
 	}
-
-	// append CSS modules as query-param
-	await promisify(require('glob'))('**/*.css', { cwd: out }).then(async cssModules => {
-		const cssData = await new Response((await new Response(cssModules.join(',')).blob()).stream().pipeThrough(new CompressionStream('gzip'))).arrayBuffer();
-		target.searchParams.set('_devCssData', Buffer.from(cssData).toString('base64'));
-	});
 
 	const emitter = new events.EventEmitter();
 	await page.exposeFunction('mocha_report', (type, data1, data2) => {
@@ -310,10 +290,8 @@ async function runTestsInBrowser(testModules, browserType) {
 	} catch (err) {
 		console.error(err);
 	}
-	if (!isDebug) {
-		server?.dispose();
-		await browser.close();
-	}
+	server.dispose();
+	await browser.close();
 
 	if (failingTests.length > 0) {
 		let res = `The followings tests are failing:\n - ${failingTests.map(({ title, message }) => `${title} (reason: ${message})`).join('\n - ')}`;
@@ -400,9 +378,7 @@ testModules.then(async modules => {
 		}
 	} catch (err) {
 		console.error(err);
-		if (!isDebug) {
-			process.exit(1);
-		}
+		process.exit(1);
 	}
 
 	// aftermath
@@ -412,9 +388,7 @@ testModules.then(async modules => {
 			console.log(msg);
 		}
 	}
-	if (!isDebug) {
-		process.exit(didFail ? 1 : 0);
-	}
+	process.exit(didFail ? 1 : 0);
 
 }).catch(err => {
 	console.error(err);
